requirements.txt
# requirement.txt

# Add your Python libraries here
pandas
sqlalchemy
python-dotenv